refactor(eve): share JSON value equality
